+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+ - [#73](https://github.com/pitest/pitclipse/pull/113) Support for JUnit 5 tests
+
 ### Removed
  - [#113](https://github.com/pitest/pitclipse/pull/113) Support for Eclipse Mars, Eclipse Neon